Modification of SAAMC - 1
I have changed the possibility to add ListOfSpecWords input because it was a variable introduced to filter words but at the same time it was in conflict with the scope of this project that is finding different communities. If we filter the words with a list of specific words we are already choosing what will possibly be a community.